add chart and lottie file2

+++ app/src/main/java/com/example/covid19_vaccine/data/domain/VaccineData.kt
@@ -3,10 +3,10 @@ package com.example.covid19_vaccine.data.domain
 data class VaccineData(
     val country: String,
     val date: String,
-    val people_vaccinated: Int,
-    val people_fully_vaccinated: Int,
-    val daily_vaccinations: Int,
+    val people_vaccinated: Double,
+    val people_fully_vaccinated: Double,
+    val daily_vaccinations: Double,
     val people_vaccinated_per_hundred: Double,
     val people_fully_vaccinated_per_hundred: Double,
-    val daily_vaccinations_per_million: Int
+    val daily_vaccinations_per_million: Double
 )
